+++ api-todo/src/config.ts
@@ -1,20 +1,34 @@
+import { getSecretFromKeyVault } from './azure/keyvault';
 // api-todo/.env
-const DATABASE_URI = process.env.DATABASE_URI;
+//const DATABASE_URI = process.env.DATABASE_URI;
 
 // ./docker-compose.yml
 const DEFAULT_MONGO_DB = 'mongodb://mongo:MongoPass@localhost:27017/';
 
-export const getConfig = (logger) => {
+function getEnvVarAsString(name: string, defaultValue: string): string {
+  return process.env[name] || defaultValue;
+}
+function getEnvVarAsBoolean(name: string, defaultValue: boolean): boolean {
+  const value = process.env[name];
+  if (value === 'true' || value === '1') {
+    return true;
+  } else if (value === 'false' || value === '0') {
+    return false;
+  } else {
+    return defaultValue;
+  }
+}
+
+export const getConfig = async (logger) => {
   const isMongoDB =
-    process.env.DATABASE_USE_MONGODB === 'true' ||
-    process.env.DATABASE_USE_MONGODB === '1'
-      ? true
-      : false;
+    getEnvVarAsString('DATABASE_USE_MONGODB', 'false') === 'true' ||
+    getEnvVarAsString('DATABASE_USE_MONGODB', 'false') === '1';
 
   logger.debug(`CONFIG: Using ${isMongoDB ? 'MongoDB' : 'in-memory'} database`);
   console.log(`CONFIG: Using ${isMongoDB ? 'MongoDB' : 'in-memory'} database`);
 
-  const dbUri = isMongoDB ? DATABASE_URI || DEFAULT_MONGO_DB : null;
+  const dbUri = isMongoDB ? await getConnectionString(logger) : null;
+
   logger.debug(`CONFIG: dbUri: ${dbUri}`);
   console.log(`CONFIG: dbUri: ${dbUri}`);
 
@@ -23,7 +37,7 @@ export const getConfig = (logger) => {
       isMongoDB,
       uri: dbUri,
       options: {
-        dbName: 'todo',
+        dbName: 'todo', // this must match the infra/app/db.bicep
       },
     },
   };
@@ -31,4 +45,28 @@ export const getConfig = (logger) => {
   console.log('CONFIG: appConfig: ', appConfig);
 
   return appConfig;
+};
+
+const getConnectionString = async (logger): Promise<string> => {
+  logger.debug(`CONFIG: getConnectionString:`);
+
+  const isCloudDb: boolean = getEnvVarAsBoolean('DATABASE_IN_CLOUD', false);
+
+  const keyVaultSecretName = getEnvVarAsString(
+    'AZURE_KEY_VAULT_COSMOSDB_CONNECTION_STRING_KEY_NAME',
+    ''
+  );
+  const keyVaultEndpoint = getEnvVarAsString('AZURE_KEY_VAULT_ENDPOINT', '');
+
+  const connectionString =
+    isCloudDb && keyVaultEndpoint && keyVaultSecretName
+      ? await getSecretFromKeyVault(
+          keyVaultEndpoint,
+          keyVaultSecretName,
+          logger
+        )
+      : DEFAULT_MONGO_DB;
+  logger.debug(`CONFIG: getConnectionString: ${connectionString}`);
+
+  return connectionString || '';
 };