+++ client-todo/src/todo/api.ts
@@ -1,1 +1,9 @@
-export const fetcher = (...args: [input: RequestInfo, init?: RequestInit]) => fetch(...args).then(res => res.json())
+export const fetcher = async (...args: [input: RequestInfo, init?: RequestInit]) => {
+    const response = await fetch(...args)
+    const { data: todos, error } = await response.json();
+
+    if (error) {
+        throw new Error(error.message);
+    }
+    return todos;
+}